# Reuse sort code in "Build dual graph"