--- a/cpp/dolfinx/common/sort.h
+++ b/cpp/dolfinx/common/sort.h
@@ -106,12 +106,11 @@ void radix_sort(xtl::span<T> array)
 /// @param[in] array The array to sort.
 /// Returns Vector of indices that sort the input array.
 template <typename T, int BITS = 8>
-std::vector<std::int32_t> argsort_radix(const xtl::span<const T>& array)
+void argsort_radix(const xtl::span<const T>& array,
+                   xtl::span<std::int32_t> perm)
 {
-  std::vector<std::int32_t> perm1(array.size());
-  std::iota(perm1.begin(), perm1.end(), 0);
   if (array.size() <= 1)
-    return perm1;
+    return;
 
   const auto [min, max] = std::minmax_element(array.begin(), array.end());
   T range = *max - *min + 1;
@@ -134,8 +133,8 @@ std::vector<std::int32_t> argsort_radix(const xtl::span<const T>& array)
   std::array<std::int32_t, bucket_size> counter;
   std::array<std::int32_t, bucket_size + 1> offset;
 
-  std::vector<std::int32_t> perm2 = perm1;
-  xtl::span<std::int32_t> current_perm = perm1;
+  std::vector<std::int32_t> perm2(perm.size());
+  xtl::span<std::int32_t> current_perm = perm;
   xtl::span<std::int32_t> next_perm = perm2;
   for (int i = 0; i < its; i++)
   {
@@ -170,10 +169,8 @@ std::vector<std::int32_t> argsort_radix(const xtl::span<const T>& array)
     mask_offset += BITS;
   }
 
-  if (its % 2 == 0)
-    return perm1;
-  else
-    return perm2;
+  if (its % 2 == 1)
+    std::copy(perm2.begin(), perm2.end(), perm.begin());
 }
 
 /// Returns the indices that would sort (lexicographic) a vector of bitsets.
@@ -308,19 +305,14 @@ std::vector<std::int32_t> sort_by_perm_new(const xt::xtensor<T, 2>& array)
   const int cols = array.shape(1);
   const int size = array.shape(0);
   std::vector<std::int32_t> perm(size);
-  std::vector<std::int32_t> local_perm(size);
   std::iota(perm.begin(), perm.end(), 0);
 
   for (int i = 0; i < cols; i++)
   {
     int col = cols - 1 - i;
-    xt::xtensor<std::int32_t, 1> column = xt::view(array, xt::keep(perm), col);
-    local_perm = argsort_radix(xtl::span<const std::int32_t>(column));
-
-    // Swap to avoid temp
-    std::swap(perm, local_perm);
-    for (int j = 0; j < size; j++)
-      perm[j] = local_perm[perm[j]];
+    xt::xtensor<std::int32_t, 1> column = xt::view(array, xt::all(), col);
+    argsort_radix<std::int32_t, BITS>(xtl::span<const std::int32_t>(column),
+                                      perm);
   }
 
   return perm;