--- a/cpp/dolfinx/mesh/graphbuild.cpp
+++ b/cpp/dolfinx/mesh/graphbuild.cpp
@@ -314,10 +314,10 @@ mesh::build_local_dual_graph(const xtl::span<const std::int64_t>& cell_vertices,
             xt::xtensor<std::int64_t, 2>({0, 0})};
   }
 
-  common::Timer t0("~0 unique identifier");
   // Give each global vertex a local identifier
-  auto perm = dolfinx::argsort_radix<std::int64_t, 16>(cell_vertices);
-  t0.stop();
+  std::vector<std::int32_t> perm(cell_vertices.size());
+  std::iota(perm.begin(), perm.end(), 0);
+  dolfinx::argsort_radix<std::int64_t, 16>(cell_vertices, perm);
 
   std::vector<std::int32_t> local_vertices(cell_vertices.size());
   std::int32_t id = 0;
@@ -334,7 +334,6 @@ mesh::build_local_dual_graph(const xtl::span<const std::int64_t>& cell_vertices,
   for (std::size_t i = 1; i < local_vertices.size(); i++)
     local_to_global[local_vertices[i]] = cell_vertices[i];
 
-  common::Timer t1("~1 count number of cells");
   // Count number of cells of each type, based on the number of vertices
   // in each cell, covering interval(2) through to hex(8)
   std::array<int, 9> count;
@@ -396,9 +395,7 @@ mesh::build_local_dual_graph(const xtl::span<const std::int64_t>& cell_vertices,
   default:
     throw std::runtime_error("Invalid tdim");
   }
-  t1.stop();
 
-  common::Timer t2("~2 list facets");
   // List of facets and associated cells
   xt::xtensor<std::int32_t, 2> facets
       = xt::empty<std::int32_t>({num_facets, num_facet_vertices});
@@ -438,7 +435,6 @@ mesh::build_local_dual_graph(const xtl::span<const std::int64_t>& cell_vertices,
   }
 
   assert(counter == (int)facets.shape(0));
-  t2.stop();
 
   // Sort facets by lexicographic order of vertices
   std::vector<std::int32_t> facet_perm = dolfinx::sort_by_perm_new(facets);