--- a/cpp/dolfinx/mesh/topologycomputation.cpp
+++ b/cpp/dolfinx/mesh/topologycomputation.cpp
@@ -440,7 +440,7 @@ compute_entities_by_key_matching(
   }
 
   // Sort the list and label uniquely
-  std::vector<std::int32_t> sort_order
+  const std::vector<std::int32_t> sort_order
       = dolfinx::sort_by_perm(entity_list_sorted);
 
   std::vector<std::int32_t> entity_index(entity_list.shape(0), 0);