- Avoid using intermediate "std::bitset" data structure.
- For lexicographic sort, sort one column of the array at a time, starting from the rightmost column.
- Work with std::int32_t in `build_local_dual_graph`: more efficient in non-comparison sort algorithms.


Example of achieved speedup using `Performance Tests`:
```
mpirun -n 6 ./dolfinx-scaling-test --ndofs 300000
```

This:
```
Compute local part of mesh dual graph                          |     4  0.684040  2.736162
ZZZ Create Mesh                                                |     1  6.757732  6.757732
ZZZ Create mesh entity permutations                            |     1  3.881517  3.881517
```

Main branch:
```
Compute local part of mesh dual graph                          |     4  0.980119  3.920477
ZZZ Create Mesh                                                |     1  8.116145  8.116145
ZZZ Create mesh entity permutations                            |     1  4.784420  4.784420
```